+++ comm/senderoscxy.cpp
@@ -29,11 +29,11 @@ SenderOscXY::SenderOscXY(RC1 *rc1)
     y=0;
 }
 
-void SenderOscXY::noteOn(int chan, int voiceId, double fr, int midinote, int pitch, int vel)
+void SenderOscXY::noteOn(int chan, int voiceId, float fr, int midinote, int pitch, int vel)
 {
 }
 
-void SenderOscXY::noteOff(int chan, int voiceId, double f, int midinote, int pitch, int vel)
+void SenderOscXY::noteOff(int chan, int voiceId, float f, int midinote, int pitch, int vel)
 {
 }
 
@@ -47,7 +47,7 @@ void SenderOscXY::pc(int chan, int v1)
 
 }
 
-void SenderOscXY::cc(int chan, int voiceId, int cc, double v1)
+void SenderOscXY::cc(int chan, int voiceId, int cc, float v1)
 {
 
     if(cc==1) x=v1;
